Fix #777: incorrect descriptions for failed generated tests
The final report was looking at the original function object to
determine the name, but by that time it has been modified and no longer
represents the test that failed.  This caches the description, so the
report summary can get the correct name.

wip: should we just pass in the description directly?  Seems like we
should.