Change how we handle KeyboardInterrupt.
8fbf666ee364ea0828fbc76ea137e274b1adb83c introduced some extra handling
for KeyboardInterrupt so that Nose would still print the summary of
errors.  Unfortunately, it also ended up swallowing the exception and
then the test suite would look like it passed when it should have been
marked a failure.

This takes another approach, and records the KeyboardInterrupt as an
error and prints the summary but it also marks the result up saying that
we should stop the tests.  This allows the tests to end, but show a
failure.